refactor: 프로젝트 패키지 구조 변경
- 클래스의 역할에 맞게 패키지 이동
