+++ src/main/java/lotto/Application.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import lotto.controller.LottoManager;
+
 public class Application {
     public static void main(String[] args) {
         LottoManager lottoManager = new LottoManager();
