+++ src/main/java/lotto/domain/LottoStatistics.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 public class LottoStatistics {
