+++ src/main/java/lotto/util/ErrorMessages.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.util;
 
 public enum ErrorMessages {
