+++ src/main/java/lotto/view/OutputView.java
@@ -1,8 +1,11 @@
-package lotto;
+package lotto.view;
 
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.LottoResult;
+import lotto.domain.Rank;
 
 public class OutputView {
