+++ src/test/java/lotto/LottoStatisticsTest.java
@@ -3,6 +3,9 @@ package lotto;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 import java.util.List;
+import lotto.domain.Lotto;
+import lotto.domain.LottoResult;
+import lotto.domain.LottoStatistics;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
